requirements.txt
flask==2.2.5
gunicorn==21.2.0
sqlalchemy
pytest
request
